refactor fetching function with useStates to global level using useContext
